+++ beets/test/helper.py
@@ -34,11 +34,10 @@ import sys
 import unittest
 from contextlib import contextmanager
 from dataclasses import dataclass
-from enum import Enum
 from functools import cache, cached_property
 from pathlib import Path
 from tempfile import gettempdir, mkdtemp, mkstemp
-from typing import TYPE_CHECKING, Any, ClassVar
+from typing import TYPE_CHECKING, Any, ClassVar, Literal
 from unittest.mock import Mock, patch
 
 import pytest
@@ -46,7 +45,7 @@ from mediafile import Image, MediaFile
 
 import beets
 import beets.plugins
-from beets import importer, logging, util
+from beets import importer, util
 from beets.autotag import AlbumInfo, TrackInfo
 from beets.importer import ImportSession
 from beets.library import Item, Library
@@ -60,29 +59,12 @@ from beets.util import (
 )
 
 if TYPE_CHECKING:
+    from types import TracebackType
+
     from requests_mock.mocker import Mocker
+    from typing_extensions import Self
 
 RUNNING_IN_CI = os.environ.get("GITHUB_ACTIONS") == "true"
-
-
-class LogCapture(logging.Handler):
-    def __init__(self):
-        logging.Handler.__init__(self)
-        self.messages = []
-
-    def emit(self, record):
-        self.messages.append(str(record.msg))
-
-
-@contextmanager
-def capture_log(logger="beets"):
-    capture = LogCapture()
-    log = logging.getLogger(logger)
-    log.addHandler(capture)
-    try:
-        yield capture.messages
-    finally:
-        log.removeHandler(capture)
 
 
 def has_program(cmd, args=["--version"]):
@@ -170,12 +152,32 @@ class IOMixin(RunMixin):
 class TestHelper(RunMixin, ConfigMixin):
     """Helper mixin for high-level cli and plugin tests.
 
-    This mixin provides methods to isolate beets' global state provide
-    fixtures.
+    This mixin provides methods to isolate beets' global state.
+
+    You may use it as a context manager in pytest fixtures in order to setup
+    tests at a class or module level. See ``module_helper`` and ``class_helper``
+    fixtures, for example.
     """
 
+    request: pytest.FixtureRequest
+
+    def __enter__(self) -> Self:
+        self.setup_beets()
+        return self
+
+    def __exit__(
+        self,
+        exc_type: type[BaseException] | None,
+        exc: BaseException | None,
+        tb: TracebackType | None,
+    ) -> Literal[False]:
+        self.teardown_beets()
+        # return False/None to propagate exceptions
+        return False
+
     @pytest.fixture(autouse=True)
-    def setup(self):
+    def setup(self, request: pytest.FixtureRequest):
+        self.request = request
         self.setup_beets()
         try:
             yield
@@ -513,7 +515,7 @@ class ImportHelper(TestHelper):
     autotagging library and several assertions for the library.
     """
 
-    default_import_config: ClassVar[dict[str, bool]] = {
+    default_import_config: ClassVar[dict[str, Any]] = {
         "autotag": True,
         "copy": True,
         "hardlink": False,
@@ -645,7 +647,6 @@ class ImportSessionFixture(ImportSession):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._choices = []
-        self._resolutions = []
 
     default_choice = importer.Action.APPLY
 
@@ -668,23 +669,6 @@ class ImportSessionFixture(ImportSession):
         return choice
 
     choose_item = choose_match
-
-    Resolution = Enum("Resolution", "REMOVE SKIP KEEPBOTH MERGE")
-
-    default_resolution = "REMOVE"
-
-    def resolve_duplicate(self, task, found_duplicates):
-        try:
-            res = self._resolutions.pop(0)
-        except IndexError:
-            res = self.default_resolution
-
-        if res == self.Resolution.SKIP:
-            task.set_choice(importer.Action.SKIP)
-        elif res == self.Resolution.REMOVE:
-            task.should_remove_duplicates = True
-        elif res == self.Resolution.MERGE:
-            task.should_merge_duplicates = True
 
 
 class TerminalImportSessionFixture(TerminalImportSession):
@@ -739,7 +723,7 @@ class TerminalImportMixin(IOMixin, ImportHelper):
             self.lib,
             loghandler=None,
             query=None,
-            io=self.io,
+            io=self.request.getfixturevalue("io"),
             paths=[import_dir],
         )
 
@@ -837,13 +821,20 @@ class AutotagStub:
         )
 
 
-class AutotagImportTestCase(ImportHelper, BeetsTestCase):
+class AutotagImportHelper(ImportHelper):
     matching = AutotagStub.IDENT
 
-    def setUp(self):
-        super().setUp()
+    def setup_beets(self):
+        super().setup_beets()
         self.matcher = AutotagStub(self.matching).install()
-        self.addCleanup(self.matcher.restore)
+
+    def teardown_beets(self):
+        self.matcher.restore()
+        super().teardown_beets()
+
+
+class AutotagImportTestCase(AutotagImportHelper, BeetsTestCase):
+    """DEPRECATED: Use AutotagImportHelper instead."""
 
 
 @dataclass(slots=True)