+++ test/test_importer.py
@@ -43,14 +43,13 @@ from beets.test.helper import (
     NEEDS_FFPROBE,
     NEEDS_REFLINK,
     AsIsImporterMixin,
+    AutotagImportHelper,
     AutotagImportTestCase,
     AutotagStub,
     BeetsTestCase,
     ImportHelper,
-    IOMixin,
     PluginMixin,
     TestHelper,
-    capture_log,
     has_program,
 )
 from beets.util import bytestring_path, displayable_path, syspath
@@ -402,7 +401,7 @@ class TestImportFormat(ImportHelper):
         self.importer.run()
         assert self.lib.items().get().path.endswith(b".mp3")
 
-    def test_recognize_format_already_exist(self):
+    def test_recognize_format_already_exist(self, caplog):
         resource_path = os.path.join(_common.RSRC, b"no_ext")
         temp_resource_path = os.path.join(self.temp_dir, b"no_ext")
         util.copy(resource_path, temp_resource_path)
@@ -410,9 +409,12 @@ class TestImportFormat(ImportHelper):
         util.copy(temp_resource_path, new_path)
         self.setup_importer(autotag=False)
         self.importer.paths = [temp_resource_path]
-        with capture_log() as logs:
+        with caplog.at_level("DEBUG"):
             self.importer.run()
-        assert "Import file with matching format to original target" in logs
+        assert (
+            "Import file with matching format to original target"
+            in caplog.messages
+        )
         assert self.lib.items().get().path.endswith(b".mp3")
 
     def test_recognize_format_not_music(self):
@@ -443,11 +445,11 @@ class TestImportFormat(ImportHelper):
         assert Path(os.path.join(self.temp_dir_path, "no_ext")).exists()
 
 
-class ImportTest(PathsMixin, AutotagImportTestCase):
+class TestImport(PathsMixin, AutotagImportHelper):
     """Test APPLY, ASIS and SKIP choices."""
 
-    def setUp(self):
-        super().setUp()
+    def setup_beets(self):
+        super().setup_beets()
         self.prepare_album_for_import(1)
         self.setup_importer()
 
@@ -536,26 +538,26 @@ class ImportTest(PathsMixin, AutotagImportTestCase):
         assert len(self.lib.items()) == 1
 
     @NEEDS_FFPROBE
-    def test_empty_directory_warning(self):
+    def test_empty_directory_warning(self, caplog):
         import_dir = os.path.join(self.temp_dir, b"empty")
         self.touch(b"non-audio", dir_=import_dir)
         self.setup_importer(import_dir=import_dir)
-        with capture_log() as logs:
+        with caplog.at_level("DEBUG"):
             self.importer.run()
 
         import_dir = displayable_path(import_dir)
-        assert f"No files imported from {import_dir}" in logs
+        assert f"No files imported from {import_dir}" in caplog.messages
 
     @NEEDS_FFPROBE
-    def test_empty_directory_singleton_warning(self):
+    def test_empty_directory_singleton_warning(self, caplog):
         import_dir = os.path.join(self.temp_dir, b"empty")
         self.touch(b"non-audio", dir_=import_dir)
         self.setup_singleton_importer(import_dir=import_dir)
-        with capture_log() as logs:
+        with caplog.at_level("DEBUG"):
             self.importer.run()
 
         import_dir = displayable_path(import_dir)
-        assert f"No files imported from {import_dir}" in logs
+        assert f"No files imported from {import_dir}" in caplog.messages
 
     def test_asis_no_data_source(self):
         assert self.lib.items().get() is None
@@ -1047,7 +1049,7 @@ class TestImportDuplicateAlbum(PluginMixin, ImportHelper):
         assert item.title == "t\xeftle 0"
         assert item.filepath.exists()
 
-        self.importer.default_resolution = self.importer.Resolution.REMOVE
+        self.config["import"]["duplicate_action"] = "remove"
         self.importer.run()
 
         assert not item.filepath.exists()
@@ -1065,7 +1067,7 @@ class TestImportDuplicateAlbum(PluginMixin, ImportHelper):
 
         assert old_artpath.exists()
 
-        self.importer.default_resolution = self.importer.Resolution.REMOVE
+        self.config["import"]["duplicate_action"] = "remove"
         self.importer.run()
 
         assert not old_artpath.exists()
@@ -1091,7 +1093,7 @@ class TestImportDuplicateAlbum(PluginMixin, ImportHelper):
         import_file.title = "new title"
         import_file.save()
 
-        self.importer.default_resolution = self.importer.Resolution.REMOVE
+        self.config["import"]["duplicate_action"] = "remove"
         self.importer.run()
 
         # Old duplicate should be removed, new one imported
@@ -1101,7 +1103,7 @@ class TestImportDuplicateAlbum(PluginMixin, ImportHelper):
         assert self.lib.items().get().title == "new title"
 
     def test_keep_duplicate_album(self):
-        self.importer.default_resolution = self.importer.Resolution.KEEPBOTH
+        self.config["import"]["duplicate_action"] = "keep"
         self.importer.run()
 
         assert len(self.lib.albums()) == 2
@@ -1111,7 +1113,7 @@ class TestImportDuplicateAlbum(PluginMixin, ImportHelper):
         item = self.lib.items().get()
         assert item.title == "t\xeftle 0"
 
-        self.importer.default_resolution = self.importer.Resolution.SKIP
+        self.config["import"]["duplicate_action"] = "skip"
         self.importer.run()
 
         assert len(self.lib.albums()) == 1
@@ -1120,7 +1122,7 @@ class TestImportDuplicateAlbum(PluginMixin, ImportHelper):
         assert item.title == "t\xeftle 0"
 
     def test_merge_duplicate_album(self):
-        self.importer.default_resolution = self.importer.Resolution.MERGE
+        self.config["import"]["duplicate_action"] = "merge"
         self.importer.run()
 
         assert len(self.lib.albums()) == 1
@@ -1141,7 +1143,7 @@ class TestImportDuplicateAlbum(PluginMixin, ImportHelper):
         import_file.title = item["title"]
         import_file.flex = "different"
 
-        self.importer.default_resolution = self.importer.Resolution.SKIP
+        self.config["import"]["duplicate_action"] = "skip"
         self.importer.run()
 
         assert len(self.lib.albums()) == 2
@@ -1182,7 +1184,7 @@ class TestImportDuplicateAlbumThreaded(PluginMixin, ImportHelper):
 
     def test_merge_duplicate_album_threaded(self):
         self.config["threaded"] = True
-        self.importer.default_resolution = self.importer.Resolution.MERGE
+        self.config["import"]["duplicate_action"] = "merge"
         self.importer.run()
 
         assert len(self.lib.albums()) == 1
@@ -1226,7 +1228,7 @@ class TestImportDuplicateSingleton(ImportHelper):
         assert item.mb_trackid == "old trackid"
         assert item.filepath.exists()
 
-        self.importer.default_resolution = self.importer.Resolution.REMOVE
+        self.config["import"]["duplicate_action"] = "remove"
         self.importer.run()
 
         assert not item.filepath.exists()
@@ -1237,7 +1239,7 @@ class TestImportDuplicateSingleton(ImportHelper):
     def test_keep_duplicate(self):
         assert len(self.lib.items()) == 1
 
-        self.importer.default_resolution = self.importer.Resolution.KEEPBOTH
+        self.config["import"]["duplicate_action"] = "keep"
         self.importer.run()
 
         assert len(self.lib.items()) == 2
@@ -1246,7 +1248,7 @@ class TestImportDuplicateSingleton(ImportHelper):
         item = self.lib.items().get()
         assert item.mb_trackid == "old trackid"
 
-        self.importer.default_resolution = self.importer.Resolution.SKIP
+        self.config["import"]["duplicate_action"] = "skip"
         self.importer.run()
 
         assert len(self.lib.items()) == 1
@@ -1260,7 +1262,7 @@ class TestImportDuplicateSingleton(ImportHelper):
         item.store()
         assert len(self.lib.items()) == 1
 
-        self.importer.default_resolution = self.importer.Resolution.SKIP
+        self.config["import"]["duplicate_action"] = "skip"
         self.importer.run()
 
         assert len(self.lib.items()) == 2
@@ -1283,7 +1285,7 @@ class TestImportDuplicateSingleton(ImportHelper):
         import_file.mb_trackid = "new trackid"
         import_file.save()
 
-        self.importer.default_resolution = self.importer.Resolution.REMOVE
+        self.config["import"]["duplicate_action"] = "remove"
         self.importer.run()
 
         # Old duplicate should be removed, new one imported
@@ -1722,45 +1724,51 @@ class ReimportTest(AutotagImportTestCase):
         assert self._album().data_source == "match_source"
 
 
-class ImportPretendTest(IOMixin, AutotagImportTestCase):
-    """Test the pretend commandline option"""
+class TestImportPretend(ImportHelper):
+    """Test the pretend commandline option."""
 
-    def setUp(self):
-        super().setUp()
+    def setup_beets(self):
+        super().setup_beets()
         self.album_track_path = self.prepare_album_for_import(1)[0]
         self.single_path = self.prepare_track_for_import(2, self.import_path)
         self.album_path = self.album_track_path.parent
 
-    def __run(self, importer):
-        with capture_log() as logs:
+    def _run(self, importer, caplog):
+        with caplog.at_level("DEBUG"):
             importer.run()
-
         assert len(self.lib.items()) == 0
         assert len(self.lib.albums()) == 0
+        return [
+            r.message
+            for r in caplog.records
+            if not r.message.startswith("Sending event:")
+        ]
 
-        return [line for line in logs if not line.startswith("Sending event:")]
-
-    def test_import_singletons_pretend(self):
-        assert self.__run(self.setup_singleton_importer(pretend=True)) == [
+    def test_import_singletons_pretend(self, caplog):
+        assert self._run(
+            self.setup_singleton_importer(pretend=True), caplog
+        ) == [
             f"Singleton: {self.single_path}",
             f"Singleton: {self.album_track_path}",
         ]
 
-    def test_import_album_pretend(self):
-        assert self.__run(self.setup_importer(pretend=True)) == [
+    def test_import_album_pretend(self, caplog):
+        assert self._run(self.setup_importer(pretend=True), caplog) == [
             f"Album: {self.import_path}",
             f"  {self.single_path}",
             f"Album: {self.album_path}",
             f"  {self.album_track_path}",
         ]
 
-    def test_import_pretend_empty(self):
+    def test_import_pretend_empty(self, caplog):
         empty_path = self.temp_dir_path / "empty"
         empty_path.mkdir()
 
         importer = self.setup_importer(pretend=True, import_dir=empty_path)
 
-        assert self.__run(importer) == [f"No files imported from {empty_path}"]
+        assert self._run(importer, caplog) == [
+            f"No files imported from {empty_path}"
+        ]
 
 
 def mocked_get_albums_by_ids(ids):